Moved some code to new ApiBase superclass
